* commit '3463360c2c0389506b005d05c2cc0a2d4ddb1dc0': (33 commits)
  Upgrade 0.40.0 to 0.41.0 OSS scripts
  DM-9610: Manage Format Lineage
  DM-9579: FactoryCQ - Remove CRED_PROXY and CRED_PORT environment variables
  - Update developer information.
  DM-9579 FactoryCQ - Remove CRED_PROXY and CRED_PORT environment variables
  DM-9579: FactoryCQ - Remove CRED_PROXY and CRED_PORT environment variables
  DM-9579 FactoryCQ - Remove CRED_PROXY and CRED_PORT environment variables
  DM-9610 change the error message
  DM-9600: Transition from S3 to Glacier without copy and delete
  DM-9610 pmd fix
  DM-9610 fix pmd
  DM-9579 FactoryCQ - Remove CRED_PROXY and CRED_PORT environment variables
  DM-9610 change name permission on update business parents
  DM-9610 correct security function name
  DM-9610 update test
  DM-9610 add check for existing of children when deleting business object format, add test and refactor existing test
  DM-9600: Transition from S3 to Glacier without copy and delete
  DM-9610 update due to story change
  DM-9595 Validate any search index
  DM-9595 Validate any search index
  ...

+++ herd-code/herd-service/src/main/java/org/finra/herd/service/impl/BusinessObjectFormatServiceImpl.java
@@ -43,6 +43,7 @@ import org.finra.herd.dao.BusinessObjectDefinitionDao;
 import org.finra.herd.dao.BusinessObjectFormatDao;
 import org.finra.herd.dao.config.DaoSpringModuleConfig;
 import org.finra.herd.model.annotation.NamespacePermission;
+import org.finra.herd.model.annotation.NamespacePermissions;
 import org.finra.herd.model.annotation.PublishJmsMessages;
 import org.finra.herd.model.api.xml.Attribute;
 import org.finra.herd.model.api.xml.AttributeDefinition;
@@ -55,6 +56,7 @@ import org.finra.herd.model.api.xml.BusinessObjectFormatDdlCollectionResponse;
 import org.finra.herd.model.api.xml.BusinessObjectFormatDdlRequest;
 import org.finra.herd.model.api.xml.BusinessObjectFormatKey;
 import org.finra.herd.model.api.xml.BusinessObjectFormatKeys;
+import org.finra.herd.model.api.xml.BusinessObjectFormatParentsUpdateRequest;
 import org.finra.herd.model.api.xml.BusinessObjectFormatUpdateRequest;
 import org.finra.herd.model.api.xml.CustomDdlKey;
 import org.finra.herd.model.api.xml.NamespacePermissionEnum;
@@ -187,13 +189,38 @@ public class BusinessObjectFormatServiceImpl implements BusinessObjectFormatServ
         Integer businessObjectFormatVersion =
             latestVersionBusinessObjectFormatEntity == null ? 0 : latestVersionBusinessObjectFormatEntity.getBusinessObjectFormatVersion() + 1;
         BusinessObjectFormatEntity newBusinessObjectFormatEntity =
-            createBusinessObjectFormatEntity(request, businessObjectDefinitionEntity, fileTypeEntity, businessObjectFormatVersion);
+            createBusinessObjectFormatEntity(request, businessObjectDefinitionEntity, fileTypeEntity, businessObjectFormatVersion, null);
         //latest version format is the descriptive format for the bdef, update the bdef descriptive format to the newly created one
         if (latestVersionBusinessObjectFormatEntity != null &&
             latestVersionBusinessObjectFormatEntity.equals(businessObjectDefinitionEntity.getDescriptiveBusinessObjectFormat()))
         {
             businessObjectDefinitionEntity.setDescriptiveBusinessObjectFormat(newBusinessObjectFormatEntity);
             businessObjectDefinitionDao.saveAndRefresh(businessObjectDefinitionEntity);
+        }
+        // latest version business object exists, update its parents and children
+        if (latestVersionBusinessObjectFormatEntity != null)
+        {
+            //for each of the previous version's  child, remove the parent link to the previous version and set the parent to the new version
+            for (BusinessObjectFormatEntity childFormatEntity : latestVersionBusinessObjectFormatEntity.getBusinessObjectFormatChildren())
+            {
+                childFormatEntity.getBusinessObjectFormatParents().remove(latestVersionBusinessObjectFormatEntity);
+                childFormatEntity.getBusinessObjectFormatParents().add(newBusinessObjectFormatEntity);
+                newBusinessObjectFormatEntity.getBusinessObjectFormatChildren().add(childFormatEntity);
+                businessObjectFormatDao.saveAndRefresh(childFormatEntity);
+            }
+            //for each of the previous version's parent, remove the child link to the previous version and set the child link to the new version
+            for (BusinessObjectFormatEntity parentFormatEntity : latestVersionBusinessObjectFormatEntity.getBusinessObjectFormatParents())
+            {
+                parentFormatEntity.getBusinessObjectFormatChildren().remove(latestVersionBusinessObjectFormatEntity);
+                parentFormatEntity.getBusinessObjectFormatChildren().add(newBusinessObjectFormatEntity);
+                newBusinessObjectFormatEntity.getBusinessObjectFormatParents().add(parentFormatEntity);
+                businessObjectFormatDao.saveAndRefresh(parentFormatEntity);
+            }
+
+            //mark the latest version business object format
+            latestVersionBusinessObjectFormatEntity.setBusinessObjectFormatParents(null);
+            latestVersionBusinessObjectFormatEntity.setBusinessObjectFormatChildren(null);
+            businessObjectFormatDao.saveAndRefresh(latestVersionBusinessObjectFormatEntity);
         }
 
         // Notify the search index that a business object definition must be updated.
@@ -362,8 +389,15 @@ public class BusinessObjectFormatServiceImpl implements BusinessObjectFormatServ
         // Retrieve and ensure that a business object format exists.
         BusinessObjectFormatEntity businessObjectFormatEntity = businessObjectFormatDaoHelper.getBusinessObjectFormatEntity(businessObjectFormatKey);
 
+        boolean checkLatestVersion = false;
+        //need to check latest version if the format key does not have the version
+        if (businessObjectFormatKey.getBusinessObjectFormatVersion() != null)
+        {
+            checkLatestVersion = true;
+        }
+
         // Create and return the business object format object from the persisted entity.
-        return businessObjectFormatHelper.createBusinessObjectFormatFromEntity(businessObjectFormatEntity);
+        return businessObjectFormatHelper.createBusinessObjectFormatFromEntity(businessObjectFormatEntity, checkLatestVersion);
     }
 
     /**
@@ -395,9 +429,25 @@ public class BusinessObjectFormatServiceImpl implements BusinessObjectFormatServ
                     businessObjectFormatHelper.businessObjectFormatEntityAltKeyToString(businessObjectFormatEntity)));
         }
 
+        if (!businessObjectFormatEntity.getBusinessObjectFormatChildren().isEmpty())
+        {
+            throw new IllegalArgumentException(String
+                .format("Can not delete a business object format that has children associated with it. Business object format: {%s}",
+                    businessObjectFormatHelper.businessObjectFormatEntityAltKeyToString(businessObjectFormatEntity)));
+        }
+
+        // Create and return the business object format object from the deleted entity.
+        BusinessObjectFormat deletedBusinessObjectFormat = businessObjectFormatHelper.createBusinessObjectFormatFromEntity(businessObjectFormatEntity);
+
+        // Check if business object format being deleted is used as a descriptive format.
+        if (businessObjectFormatEntity.equals(businessObjectFormatEntity.getBusinessObjectDefinition().getDescriptiveBusinessObjectFormat()))
+        {
+            businessObjectFormatEntity.getBusinessObjectDefinition().setDescriptiveBusinessObjectFormat(null);
+            businessObjectDefinitionDao.saveAndRefresh(businessObjectFormatEntity.getBusinessObjectDefinition());
+        }
+
         // Delete this business object format.
         businessObjectFormatDao.delete(businessObjectFormatEntity);
-
         // If this business object format version is the latest, set the latest flag on the previous version of this object format, if it exists.
         if (businessObjectFormatEntity.getLatestVersion())
         {
@@ -422,8 +472,7 @@ public class BusinessObjectFormatServiceImpl implements BusinessObjectFormatServ
         // Notify the search index that a business object definition must be updated.
         searchIndexUpdateHelper.modifyBusinessObjectDefinitionInSearchIndex(businessObjectDefinitionEntity, SEARCH_INDEX_UPDATE_TYPE_UPDATE);
 
-        // Create and return the business object format object from the deleted entity.
-        return businessObjectFormatHelper.createBusinessObjectFormatFromEntity(businessObjectFormatEntity);
+        return deletedBusinessObjectFormat;
     }
 
     /**
@@ -481,6 +530,58 @@ public class BusinessObjectFormatServiceImpl implements BusinessObjectFormatServ
     {
         return generateBusinessObjectFormatDdlCollectionImpl(request);
     }
+
+    /**
+     * Update business object format parents
+     *
+     * @param businessObjectFormatKey business object format key
+     * @param businessObjectFormatParentsUpdateRequest business object format parents update request
+     *
+     * @return business object format
+     */
+    @NamespacePermissions({@NamespacePermission(fields = "#businessObjectFormatKey.namespace", permissions = NamespacePermissionEnum.WRITE),
+        @NamespacePermission(fields = "#businessObjectFormatParentsUpdateRequest?.businessObjectFormatParents?.![namespace]",
+            permissions = NamespacePermissionEnum.READ)})
+    @Override
+    public BusinessObjectFormat updateBusinessObjectFormatParents(BusinessObjectFormatKey businessObjectFormatKey,
+        BusinessObjectFormatParentsUpdateRequest businessObjectFormatParentsUpdateRequest)
+    {
+        Assert.notNull(businessObjectFormatParentsUpdateRequest, "A Business Object Format Parents Update Request is required.");
+
+        // Perform validation and trim the alternate key parameters.
+        businessObjectFormatHelper.validateBusinessObjectFormatKey(businessObjectFormatKey, false);
+
+        Assert.isNull(businessObjectFormatKey.getBusinessObjectFormatVersion(), "Business object format version must not be specified.");
+        // Perform validation and trim for the business object format parents
+        validateBusinessObjectFormatParents(businessObjectFormatParentsUpdateRequest.getBusinessObjectFormatParents());
+
+        // Retrieve and ensure that a business object format exists.
+        BusinessObjectFormatEntity businessObjectFormatEntity = businessObjectFormatDaoHelper.getBusinessObjectFormatEntity(businessObjectFormatKey);
+
+        // Retrieve and ensure that business object format parents exist. A set is used to ignore duplicate business object format parents.
+        Set<BusinessObjectFormatEntity> businessObjectFormatParents = new HashSet<>();
+        for (BusinessObjectFormatKey businessObjectFormatParent : businessObjectFormatParentsUpdateRequest.getBusinessObjectFormatParents())
+        {
+            // Retrieve and ensure that a business object format exists.
+            BusinessObjectFormatEntity businessObjectFormatEntityParent =
+                businessObjectFormatDaoHelper.getBusinessObjectFormatEntity(businessObjectFormatParent);
+            businessObjectFormatParents.add(businessObjectFormatEntityParent);
+        }
+
+        // Set the business object format parents.
+        businessObjectFormatEntity.setBusinessObjectFormatParents(new ArrayList<>(businessObjectFormatParents));
+
+        // Persist and refresh the entity.
+        businessObjectFormatEntity = businessObjectFormatDao.saveAndRefresh(businessObjectFormatEntity);
+
+        // Notify the search index that a business object definition must be updated.
+        searchIndexUpdateHelper
+            .modifyBusinessObjectDefinitionInSearchIndex(businessObjectFormatEntity.getBusinessObjectDefinition(), SEARCH_INDEX_UPDATE_TYPE_UPDATE);
+
+        // Create and return the business object format object from the persisted entity.
+        return businessObjectFormatHelper.createBusinessObjectFormatFromEntity(businessObjectFormatEntity);
+    }
+
 
     /**
      * Retrieves the DDL to initialize the specified type of the database system (e.g. Hive) by creating a table for the requested business object format.
@@ -628,6 +729,25 @@ public class BusinessObjectFormatServiceImpl implements BusinessObjectFormatServ
     }
 
     /**
+     * Validate the business object format parents
+     *
+     * @param businessObjectFormatParents business object format parents
+     */
+    private void validateBusinessObjectFormatParents(List<BusinessObjectFormatKey> businessObjectFormatParents)
+    {
+        // Validate parents business object format.
+        if (!CollectionUtils.isEmpty(businessObjectFormatParents))
+        {
+            for (BusinessObjectFormatKey parentBusinessObjectFormatKey : businessObjectFormatParents)
+            {
+                Assert.notNull(parentBusinessObjectFormatKey, "Parent object format must be specified.");
+                businessObjectFormatHelper.validateBusinessObjectFormatKey(parentBusinessObjectFormatKey, false);
+                Assert.isNull(parentBusinessObjectFormatKey.getBusinessObjectFormatVersion(), "Business object format version must not be specified.");
+            }
+        }
+    }
+
+    /**
      * Validates the business object format schema. This method also trims some of the schema column attributes parameters.
      *
      * @param schema the business object format schema
@@ -740,13 +860,13 @@ public class BusinessObjectFormatServiceImpl implements BusinessObjectFormatServ
         // Validate that there are no changes to the delimiter character, which is a an optional parameter.
         // Please note that null and an empty string values are both stored in the database as NULL.
         Assert.isTrue(oldSchema.getDelimiter() == null ? newSchema.getDelimiter() == null || newSchema.getDelimiter().isEmpty() :
-                oldSchema.getDelimiter().equals(newSchema.getDelimiter()),
+            oldSchema.getDelimiter().equals(newSchema.getDelimiter()),
             String.format("%s New format version delimiter character does not match to the previous format version delimiter character.", mainErrorMessage));
 
         // Validate that there are no changes to the escape character, which is a an optional parameter.
         // Please note that null and an empty string values are both stored in the database as NULL.
         Assert.isTrue(oldSchema.getEscapeCharacter() == null ? newSchema.getEscapeCharacter() == null || newSchema.getEscapeCharacter().isEmpty() :
-                oldSchema.getEscapeCharacter().equals(newSchema.getEscapeCharacter()),
+            oldSchema.getEscapeCharacter().equals(newSchema.getEscapeCharacter()),
             String.format("%s New format version escape character does not match to the previous format version escape character.", mainErrorMessage));
 
         // Validate that there are no non-additive changes to partition columns.
@@ -756,7 +876,7 @@ public class BusinessObjectFormatServiceImpl implements BusinessObjectFormatServ
         // Validate that there are no non-additive changes to the previous format version regular columns.
         // No null check is needed on schema columns, since at least one column is required if format schema is specified.
         Assert.isTrue((oldSchema.getColumns().size() <= newSchema.getColumns().size()) &&
-                validateNewSchemaColumnsAreAdditiveToOldSchemaColumns(newSchema.getColumns().subList(0, oldSchema.getColumns().size()), oldSchema.getColumns()),
+            validateNewSchemaColumnsAreAdditiveToOldSchemaColumns(newSchema.getColumns().subList(0, oldSchema.getColumns().size()), oldSchema.getColumns()),
             String.format("%s Non-additive changes detected to the previously defined regular (non-partitioning) columns.", mainErrorMessage));
     }
 
@@ -923,11 +1043,13 @@ public class BusinessObjectFormatServiceImpl implements BusinessObjectFormatServ
      * Creates and persists a new business object format entity from the request information.
      *
      * @param request the request.
+     * @param businessObjectFormatEntityParents parent business object format entity
      *
      * @return the newly created business object format entity.
      */
     private BusinessObjectFormatEntity createBusinessObjectFormatEntity(BusinessObjectFormatCreateRequest request,
-        BusinessObjectDefinitionEntity businessObjectDefinitionEntity, FileTypeEntity fileTypeEntity, Integer businessObjectFormatVersion)
+        BusinessObjectDefinitionEntity businessObjectDefinitionEntity, FileTypeEntity fileTypeEntity, Integer businessObjectFormatVersion,
+        List<BusinessObjectFormatEntity> businessObjectFormatEntityParents)
     {
         BusinessObjectFormatEntity businessObjectFormatEntity = new BusinessObjectFormatEntity();
         businessObjectFormatEntity.setBusinessObjectDefinition(businessObjectDefinitionEntity);
@@ -953,10 +1075,14 @@ public class BusinessObjectFormatServiceImpl implements BusinessObjectFormatServ
             }
         }
 
+
         businessObjectFormatEntity.setAttributeDefinitions(createAttributeDefinitionEntities(request.getAttributeDefinitions(), businessObjectFormatEntity));
 
         // Add optional schema information.
         populateBusinessObjectFormatSchema(businessObjectFormatEntity, request.getSchema());
+
+        //set the parents
+        businessObjectFormatEntity.setBusinessObjectFormatParents(businessObjectFormatEntityParents);
 
         // Persist and return the new entity.
         return businessObjectFormatDao.saveAndRefresh(businessObjectFormatEntity);
